move GuiResearchBrowser mixins into their own classes instead of one giant class that is applied if any are true, also add additional creative functionality
